pyproject.toml
[project]
name = "trufas"
version = "0.1.0"
description = ""
authors = [
    {name = "Leticia", email = "leticiaborgess.oficial@gmail.com"},
]
dependencies = [
    "django>=4.2.5",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
    "setuptools>=68.2.2",
    "python-magic>=0.4.27",
    "Pillow>=10.0.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
